requirements.txt
# Core dependencies
gradio==5.49.1
typer>=0.9.0
rich>=13.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# LLM Integration
huggingface-hub>=0.20.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Optional: Local inference
# ollama>=0.3.0

# Sandbox execution
docker>=7.0.0

# Dev dependencies
pytest>=7.0.0
black>=23.0.0
mypy>=1.0.0
pytest-asyncio
openai>=1.0.0
prompt_toolkit>=3.0.0
textual>=6.0.0
psutil>=5.9.0
